fix(vitest): cover the isolated test-db factories with the speed pragmas
createIsolatedTestDb passes __smrtSkipVitestSchemaPreparation internally,
so its handles bypassed the setup mock's pragma application and kept full
durability (codex review finding). The helper now lives in
sqlite-schema-template.ts — side-effect-free, so test-db.ts can share it
without importing the setup file's vi.mock — and the factory applies it
after opening the base connection. The raw getDatabase + skip-flag path
remains fully vanilla as the documented durability escape hatch.
